+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Adding `s::enqueueNoUiSliderLibs()`.
 - Updating to latest WebSharks Core, which now includes Stripe integration.
 - Updating to latest Semantic UI build, now a custom build that comes with the WebSharks Core.
+- Adding support for inline scripts via `wp_add_inline_script()`.
 
 ## v170311.43193